+++ R/frs_watershed_at_measure.R
@@ -22,7 +22,7 @@
 #'
 #' @return An `sf` data frame with a single polygon geometry.
 #'
-#' @family network
+#' @family watershed
 #'
 #' @export
 #'